requirements.txt
# required packages

requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv
transformers
langchain